refactor: granualized models
