+++ patients/models/__init__.py
@@ -1,0 +1,4 @@
+from .patient import Patient
+from .doctor import Doctor
+from .appointment import Appointment
+from .invoice import Invoice
